Replace old code with the new chat bot
